refactor: Обновит импорты в компонентах. Удалит старый файл api.ts
