+++ src/pages/PostDetailPage/ui/PostDetailPage.tsx
@@ -2,9 +2,10 @@ import { useParams } from "react-router-dom";
 import { Error } from "@/shared/ui/Error/Error";
 import { withLoading } from "@/shared/lib/hoc/withLoading";
 import { PostContent } from "@/entities/post/ui/PostContent";
+import { useGetPostByIdQuery } from "@/entities/post/api/postApi";
 import { CommentListDefault } from "@/widgets/CommentList/ui/CommentList";
 import { LoadingSpinner } from "@/shared/ui/LoadingSpinner/LoadingSpinner";
-import { useGetCommentsByPostIdQuery, useGetPostByIdQuery } from "@/shared/api/api";
+import { useGetCommentsByPostIdQuery } from "@/entities/comment/api/commentsApi";
 
 const CommentListWithLoading = withLoading(CommentListDefault);
 
@@ -15,8 +16,8 @@ export const PostDetailPage = () => {
     return <Error errorMessage="Некорректный ID" />;
   }
 
-  const {data: post, isLoading, isError, refetch} = useGetPostByIdQuery(id);
-  const {data: comments, isLoading: isCommentsLoading, isError: isCommentsError, refetch: refetchComments} = useGetCommentsByPostIdQuery(id);
+  const {data: post, isLoading, isError, refetch} = useGetPostByIdQuery(Number(id));
+  const {data: comments, isLoading: isCommentsLoading, isError: isCommentsError, refetch: refetchComments} = useGetCommentsByPostIdQuery(Number(id));
 
   if(isLoading) {
       return <LoadingSpinner />;
